+++ day11/src/Utils/Monkey.java
@@ -12,12 +12,12 @@ import java.util.NoSuchElementException;
 public class Monkey {
     private final String name;
     private final String operation;
-    private final Integer test;
+    private final Long test;
     private final String testTrue;
     private final String testFalse;
     private List<Long> inventory;
     @Setter
-    private Integer inspections;
+    private Long inspections;
 
     public static Monkey getMonkeyByName(String name, List<Monkey> monkeyList) {
         return monkeyList.stream()
